requirements.txt
pandas
numpy
setuptools
kagglehub
torch
opencv-python-headless
torchvision
tensorboard
dvc
dvc-gs
fastapi[all]
uvicorn
Pillow